--- a/.github/workflows/workflow.yml
+++ b/.github/workflows/workflow.yml
@@ -37,6 +37,6 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - name: golangci-lint
-        uses: golangci/golangci-lint-action@v2
+        uses: golangci/golangci-lint-action@v3
         with:
           version: latest