--- a/README.md
+++ b/README.md
@@ -29,6 +29,6 @@ Licensed under either of <a href="LICENSE-APACHE">Apache License, Version
 
 <sub>
 Unless you explicitly state otherwise, any contribution intentionally submitted
-for inclusion in Proteus by you, as defined in the Apache-2.0 license, shall be
+for inclusion in this package by you, as defined in the Apache-2.0 license, shall be
 dual licensed as above, without any additional terms or conditions.
 </sub>