--- a/go.mod
+++ b/go.mod
@@ -2,6 +2,6 @@ module github.com/go-playground/cache
 
 go 1.20
 
-require github.com/go-playground/pkg/v5 v5.13.0
+require github.com/go-playground/pkg/v5 v5.14.0
 
 require github.com/go-playground/assert/v2 v2.2.0