--- a/lfu/lfu.go
+++ b/lfu/lfu.go
@@ -2,6 +2,7 @@ package lfu
 
 import (
 	listext "github.com/go-playground/pkg/v5/container/list"
+	timeext "github.com/go-playground/pkg/v5/time"
 	optionext "github.com/go-playground/pkg/v5/values/option"
 	"sync"
 	"time"
@@ -24,7 +25,7 @@ func New[K comparable, V any](capacity int) *builder[K, V] {
 
 // MaxAge sets the maximum age of an entry before it should be discarded; passively.
 func (b *builder[K, V]) MaxAge(maxAge time.Duration) *builder[K, V] {
-	b.lfu.maxAge = maxAge
+	b.lfu.maxAge = int64(maxAge)
 	return b
 }
 
@@ -65,7 +66,7 @@ func (b *builder[K, V]) Build() (lru *Cache[K, V]) {
 type entry[K comparable, V any] struct {
 	key       K
 	value     V
-	ts        time.Time
+	ts        int64
 	frequency *listext.Node[frequency[K, V]]
 }
 
@@ -80,7 +81,7 @@ type Cache[K comparable, V any] struct {
 	frequencies        *listext.DoublyLinkedList[frequency[K, V]]
 	entries            map[K]*listext.Node[entry[K, V]]
 	capacity           int
-	maxAge             time.Duration
+	maxAge             int64
 	lastPercentageFull uint8
 	hitFn              func(key K, value V)
 	missFn             func(key K)
@@ -96,7 +97,7 @@ func (cache *Cache[K, V]) Set(key K, value V) {
 	if found {
 		node.Value.value = value
 		if cache.maxAge > 0 {
-			node.Value.ts = time.Now()
+			node.Value.ts = timeext.NanoTime()
 		}
 		node.Value.frequency.Value.entries.MoveToFront(node)
 	} else {
@@ -115,7 +116,7 @@ func (cache *Cache[K, V]) Set(key K, value V) {
 			frequency: freq,
 		}
 		if cache.maxAge > 0 {
-			e.ts = time.Now()
+			e.ts = timeext.NanoTime()
 		}
 		cache.entries[key] = freq.Value.entries.PushFront(e)
 		if len(cache.entries) > cache.capacity {
@@ -154,7 +155,7 @@ func (cache *Cache[K, V]) Get(key K) (result optionext.Option[V]) {
 
 	node, found := cache.entries[key]
 	if found {
-		if cache.maxAge > 0 && time.Since(node.Value.ts) > cache.maxAge {
+		if cache.maxAge > 0 && timeext.NanoTime()-node.Value.ts > cache.maxAge {
 			cache.remove(node)
 			if cache.evictFn != nil {
 				cache.evictFn(key, node.Value.value)