--- a/lfu/lfu_test.go
+++ b/lfu/lfu_test.go
@@ -284,3 +284,16 @@ func BenchmarkLFUCacheSetsOnly(b *testing.B) {
 		cache.Set(j, "b")
 	}
 }
+
+func BenchmarkLFUCacheSetGetDynamicWithEvictions(b *testing.B) {
+	cache := New[string, string](100).Build()
+
+	for i := 0; i < b.N; i++ {
+		j := strconv.Itoa(i)
+		cache.Set(j, j)
+		option := cache.Get(j)
+		if option.IsNone() || option.Unwrap() != j {
+			panic("undefined behaviour")
+		}
+	}
+}