--- a/lru/lru_test.go
+++ b/lru/lru_test.go
@@ -187,3 +187,16 @@ func BenchmarkLRUCacheSetGetDynamicWithEvictions(b *testing.B) {
 		}
 	}
 }
+
+func BenchmarkLRUCacheGetSetParallel(b *testing.B) {
+	cache := New[string, string](100).Build()
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			cache.Set("a", "b")
+			option := cache.Get("a")
+			if option.IsNone() || option.Unwrap() != "b" {
+				panic("undefined behaviour")
+			}
+		}
+	})
+}