### Changed
- Renamed LRU and LFU cache struct to Cache to prevent stutter in calling code. Now reads `lru.Cache` and `lfu.Cache`.
- Changed to use `int64` and monotonic time instead of `time.*` functions for speed and better space efficiency.

### Added
- Optimization when incrementing frequency and the found node is the only entry.

```go
benchmark                                                 old ns/op     new ns/op     delta
BenchmarkLFUCacheWithAllRegisteredFunctions-8             143           69.2          -51.61%
BenchmarkLFUCacheNoRegisteredFunctions-8                  144           66.3          -54.03%
BenchmarkLFUCacheWithAllRegisteredFunctionsNoMaxAge-8     71.6          28.6          -60.10%
BenchmarkLFUCacheNoRegisteredFunctionsNoMaxAge-8          70.1          27.7          -60.47%
BenchmarkLFUCacheGetsOnly-8                               60.0          13.4          -77.61%
BenchmarkLFUCacheSetsOnly-8                               149           147           -1.81%
BenchmarkLFUCacheSetGetDynamicWithEvictions-8             223           218           -2.29%
BenchmarkLRUCacheWithAllRegisteredFunctions-8             99.1          73.4          -25.95%
BenchmarkLRUCacheNoRegisteredFunctions-8                  96.1          70.1          -27.10%
BenchmarkLRUCacheWithAllRegisteredFunctionsNoMaxAge-8     31.0          31.4          +1.45%
BenchmarkLRUCacheNoRegisteredFunctionsNoMaxAge-8          28.0          28.2          +0.79%
BenchmarkLRUCacheGetsOnly-8                               14.9          14.9          -0.40%
BenchmarkLRUCacheSetsOnly-8                               149           145           -2.76%
BenchmarkLRUCacheSetGetDynamicWithEvictions-8             174           169           -2.65%

benchmark                                                 old allocs     new allocs     delta
BenchmarkLFUCacheWithAllRegisteredFunctions-8             2              0              -100.00%
BenchmarkLFUCacheNoRegisteredFunctions-8                  2              0              -100.00%
BenchmarkLFUCacheWithAllRegisteredFunctionsNoMaxAge-8     2              0              -100.00%
BenchmarkLFUCacheNoRegisteredFunctionsNoMaxAge-8          2              0              -100.00%
BenchmarkLFUCacheGetsOnly-8                               2              0              -100.00%
BenchmarkLFUCacheSetsOnly-8                               2              2              +0.00%
BenchmarkLFUCacheSetGetDynamicWithEvictions-8             4              4              +0.00%
BenchmarkLRUCacheWithAllRegisteredFunctions-8             0              0              +0.00%
BenchmarkLRUCacheNoRegisteredFunctions-8                  0              0              +0.00%
BenchmarkLRUCacheWithAllRegisteredFunctionsNoMaxAge-8     0              0              +0.00%
BenchmarkLRUCacheNoRegisteredFunctionsNoMaxAge-8          0              0              +0.00%
BenchmarkLRUCacheGetsOnly-8                               0              0              +0.00%
BenchmarkLRUCacheSetsOnly-8                               2              2              +0.00%
BenchmarkLRUCacheSetGetDynamicWithEvictions-8             2              2              +0.00%

benchmark                                                 old bytes     new bytes     delta
BenchmarkLFUCacheWithAllRegisteredFunctions-8             56            0             -100.00%
BenchmarkLFUCacheNoRegisteredFunctions-8                  56            0             -100.00%
BenchmarkLFUCacheWithAllRegisteredFunctionsNoMaxAge-8     56            0             -100.00%
BenchmarkLFUCacheNoRegisteredFunctionsNoMaxAge-8          56            0             -100.00%
BenchmarkLFUCacheGetsOnly-8                               56            0             -100.00%
BenchmarkLFUCacheSetsOnly-8                               97            81            -16.49%
BenchmarkLFUCacheSetGetDynamicWithEvictions-8             154           138           -10.39%
BenchmarkLRUCacheWithAllRegisteredFunctions-8             0             0             +0.00%
BenchmarkLRUCacheNoRegisteredFunctions-8                  0             0             +0.00%
BenchmarkLRUCacheWithAllRegisteredFunctionsNoMaxAge-8     0             0             +0.00%
BenchmarkLRUCacheNoRegisteredFunctionsNoMaxAge-8          0             0             +0.00%
BenchmarkLRUCacheGetsOnly-8                               0             0             +0.00%
BenchmarkLRUCacheSetsOnly-8                               97            81            -16.49%
BenchmarkLRUCacheSetGetDynamicWithEvictions-8             97            81            -16.49%
```